django/ibg/requirements.txt
django==4.0.3
asgiref==3.5.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
joblib==1.1.0
mysql-client==0.0.1
mysqlclient==2.1.0
numpy==1.22.3
pandas==1.4.1
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.0.2
scipy==1.8.0
setuptools==58.1.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
wheel==0.36.2
gunicorn==20.1.0
